+++ docpie/__init__.py
@@ -17,7 +17,7 @@ __all__ = ('docpie', 'Docpie',
 # it's not a good idea but it can avoid loop importing
 __version__ = Docpie._version
 
-__timestamp__ = 1449989766.593201  # last sumbit
+__timestamp__ = 1450027365.346833  # last sumbit
 
 logger = getLogger('docpie')